refactor app.tsx and add react-query
